-Woverloaded-virtual made an error (to avoid API misuse)
This helps prevent coding errors when an override of a method was
intended, but it has a wrong signature: then it would trigger this
warning, because the overriding method in the derived class hides all
other overloaded methods from the base class (according to the C++
standard).

An example of such strange incorrect overrides could be found in the
ugly common/rswig.h workarounds (dropped in the previous commit).