*.conf: do not supply a text for a key which has the same default value
The default value is hardcoded and therefore a translation for it is
normally provided.

Contrarily, if a user supplies his own text in the configuration,
there is no hope a translation would be found among the compiled ones.

And therefore putting a "custom" value for Synaptic::supported-text in
the packaged config file was evil: this prevented looking up the
translations for this label and prevented showing this label
translated in the UI.

Fixes: e25bc10 Fix "Package is supported" label translation
Fixes: 34d0e97 Apply synaptic-0.58-rgiconlegend-supported.patch