RPackageCache: ensure that owned allocated objects get deleted on destruction
Perhaps, this is not important in synaptic, but this way the code
looks better: it's evident that these allocated objects can't leak.

Note: RPackageCache class in synaptic is quite similar to CacheFile in apt.

FIXME: Actually, I don't like playing with these references much in
this and the following commit (and in the corresponding places in APT
source code--in the CacheFile related code): these changes don't
capture the essence anyway, namely, that one object depends on the
other (in a sequence), and the owner should care not to delete/destroy
_map while _cache is used. _map is a prerequisite for _cache; _cache
for _policy; _cache and _policy for _dcache. All of them save the
pointers/references to their prerequisites, and they are in danger of
becoming dangling if there is an error in the owning code. We could
try better to formally express this in their constructors or even
factories and allow them to be owned only by the same owner which owns
their prerequisite objects. (FileFd is not of this kind, because an
mmap persists after closing the used fd.)

I envision coupling dependent objects to the prerequisite object with
a special class (on top of unique_ptr or the like).